feat: implement static mapping extensions and refactor metric, target to double

+++ SkillTree/DTOs/CreateSkillDto.cs
@@ -5,5 +5,5 @@ public class CreateSkillDto
     public string Name { get; set; } = string.Empty;
     public string? Description { get; set; }
     public string Metric { get; set; } = string.Empty;
-    public int? Target  { get; set; }
+    public double? Target  { get; set; }
 }
